+++ packages/codefixes/test/fixtures/ts-codefixes/deleteUnmatchedParameter/passing/pass-1.ts
@@ -1,4 +1,4 @@
-/**
+  /**
    * @param {number} a
    */
 export function foo(a: number) {