+++ packages/plugins/src/index.ts
@@ -1,1 +1,10 @@
 export * from './plugins/index.js';
+export {
+  DummyPlugin,
+  Plugin,
+  PluginFactory,
+  PluginOptions,
+  PluginsRunnerContext,
+  PluginsRunner,
+} from './plugin.js';
+export type { PluginResult } from './plugin.js';