+++ packages/plugins/src/plugins/diagnostic-fix.plugin.ts
@@ -9,10 +9,12 @@ import {
   type DiagnosticWithContext,
   applyCodeFix,
 } from '@rehearsal/codefixes';
-import { PluginOptions, PluginsRunnerContext, Service, Plugin } from '@rehearsal/service';
 import { findNodeAtPosition } from '@rehearsal/ts-utils';
-import type MS from 'magic-string';
+import { PluginOptions, PluginsRunnerContext, Plugin } from '../plugin.js';
 import { getFormatCodeSettingsForFile } from '../helpers.js';
+
+import type { Service } from '@rehearsal/service';
+import type MS from 'magic-string';
 
 const require = createRequire(import.meta.url);
 // eslint-disable-next-line @typescript-eslint/no-unsafe-assignment