+++ packages/service/src/index.ts
@@ -1,14 +1,4 @@
-export {
-  DummyPlugin,
-  Plugin,
-  PluginFactory,
-  PluginOptions,
-  PluginsRunnerContext,
-  PluginsRunner,
-} from './plugin.js';
 export { RehearsalService, type Service } from './rehearsal-service.js';
 export { RehearsalServiceHost } from './rehearsal-service-host.js';
 export { GlintService, type Range, type PathUtils } from './glint-service.js';
 export * from './glint-utils.js';
-
-export type { PluginResult } from './plugin.js';